Upgrade repository import api
