+++ scm-core/src/main/java/sonia/scm/repository/spi/RemoteCommandRequest.java
@@ -21,57 +21,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.repository.spi;
 
-//~--- non-JDK imports --------------------------------------------------------
-
-import com.google.common.base.MoreObjects;
-import com.google.common.base.Objects;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 import sonia.scm.repository.Repository;
 
 import java.net.URL;
 
-//~--- JDK imports ------------------------------------------------------------
-
 /**
- *
  * @author Sebastian Sdorra
  * @since 1.31
  */
-public abstract class RemoteCommandRequest implements Resetable
-{
+@Getter
+@Setter
+@EqualsAndHashCode
+@ToString
+public abstract class RemoteCommandRequest implements Resetable {
 
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public boolean equals(Object obj)
-  {
-    if (obj == null)
-    {
-      return false;
-    }
-
-    if (getClass() != obj.getClass())
-    {
-      return false;
-    }
-
-    final RemoteCommandRequest other = (RemoteCommandRequest) obj;
-
-    return Objects.equal(remoteRepository, other.remoteRepository)
-      && Objects.equal(remoteUrl, other.remoteUrl);
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public int hashCode()
-  {
-    return Objects.hashCode(remoteRepository, remoteUrl);
-  }
+  protected Repository remoteRepository;
+  protected URL remoteUrl;
+  protected String username;
+  protected String password;
 
   /**
    * Resets the request object.
@@ -79,82 +53,10 @@ public abstract class RemoteCommandRequest implements Resetable
    * @since 1.43
    */
   @Override
-  public void reset()
-  {
+  public void reset() {
     remoteRepository = null;
     remoteUrl = null;
+    username = null;
+    password = null;
   }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public String toString()
-  {
-    //J-
-    return MoreObjects.toStringHelper(this)
-                  .add("remoteRepository", remoteRepository)
-                  .add("remoteUrl", remoteUrl)
-                  .toString();
-    //J+
-  }
-
-  //~--- set methods ----------------------------------------------------------
-
-  /**
-   * Method description
-   *
-   * @param remoteRepository
-   */
-  public void setRemoteRepository(Repository remoteRepository)
-  {
-    this.remoteRepository = remoteRepository;
-  }
-
-  /**
-   * Method description
-   *
-   *
-   * @param remoteUrl
-   *
-   * @since 1.43
-   */
-  public void setRemoteUrl(URL remoteUrl)
-  {
-    this.remoteUrl = remoteUrl;
-  }
-
-  //~--- get methods ----------------------------------------------------------
-
-  /**
-   * Method description
-   *
-   *
-   * @return
-   */
-  Repository getRemoteRepository()
-  {
-    return remoteRepository;
-  }
-
-  /**
-   * Method description
-   *
-   *
-   * @return
-   *
-   * @since 1.43
-   */
-  URL getRemoteUrl()
-  {
-    return remoteUrl;
-  }
-
-  //~--- fields ---------------------------------------------------------------
-
-  /** Field description */
-  protected Repository remoteRepository;
-
-  /** remote url */
-  protected URL remoteUrl;
 }
